teachable machine(Project)/requirements.txt
   Flask==2.3.3   Flask-CORS==4.0.0   azure-cognitiveservices-vision-customvision==3.1.0   azure-cognitiveservices-vision-computervision==0.9.0   msrest==0.7.1   Pillow==10.0.1   python-dotenv==1.0.0   tensorflow==2.15.0   Flask-SQLAlchemy==3.1.1   celery==5.3.4   redis==5.0.1